--- app/forms/supply_teacher_form.rb
@@ -1,34 +1,0 @@
-class SupplyTeacherForm < Form
-  attribute :employed_as_supply_teacher, :boolean
-
-  validates :employed_as_supply_teacher,
-    inclusion: {
-      in: [true, false],
-      message: ->(object, _) { object.i18n_errors_path("select_employed_as_supply_teacher") }
-    }
-
-  def initialize(claim:, journey:, params:)
-    super
-
-    self.employed_as_supply_teacher = permitted_params.fetch(
-      :employed_as_supply_teacher,
-      claim.eligibility.employed_as_supply_teacher
-    )
-  end
-
-  def save
-    return false unless valid?
-
-    update!({eligibility_attributes: {employed_as_supply_teacher:}})
-  end
-
-  private
-
-  def i18n_form_namespace
-    "supply_teacher"
-  end
-
-  def permitted_params
-    @permitted_params ||= params.fetch(:claim, {}).permit(:employed_as_supply_teacher)
-  end
-end
