+++ app/views/additional_payments/claims/supply_teacher.html.erb
@@ -1,13 +1,12 @@
 <% content_for(:page_title, page_title(t("#{@form.i18n_namespace}.forms.supply_teacher.questions.employed_as_supply_teacher"), journey: current_journey_routing_name, show_error: @form.errors.any?)) %>
-<% path_for_form = current_claim.persisted? ? claim_path(current_journey_routing_name) : claims_path(current_journey_routing_name) %>
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
     <%= render("shared/error_summary", instance: @form, errored_field_id_overrides: { "eligibility.employed_as_supply_teacher": "claim_eligibility_attributes_employed_as_supply_teacher_true" }) if @form.errors.any? %>
 
-    <%= form_for @form, url: path_for_form, method: :patch do |form| %>
+    <%= form_for @form, url: claim_path(current_journey_routing_name) do |form| %>
       <%= form_group_tag @form do %>
-        <%= hidden_field_tag :employed_as_supply_teacher, params[:employed_as_supply_teacher] %>
+        <%= form.hidden_field :employed_as_supply_teacher %>
 
         <fieldset class="govuk-fieldset">
           <legend class="govuk-fieldset__legend <%= fieldset_legend_css_class_for_journey(journey) %>">
